requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.40.0
sentencepiece>=0.2.0
aiofiles>=23.2.0
pydantic>=2.5.0
numpy>=1.24.0,<2.0.0
ffmpeg-python>=0.2.0
websockets>=12.0
llama-cpp-python>=0.2.0
streamlink>=6.0.0
yt-dlp>=2024.0.0
requests>=2.31.0
jinja2>=3.1.0
ollama>=0.1.0
